refactor: move back button to the tab menu
