+++ src/renderer/components/App/index.js
@@ -1,10 +1,8 @@
-import AppBackButton from './AppBackButton'
 import AppSidemenu from './AppSidemenu/AppSidemenu'
 import AppFooter from './AppFooter'
 import AppIntro from './AppIntro/AppIntro'
 
 export {
-  AppBackButton,
   AppSidemenu,
   AppFooter,
   AppIntro
