+++ src/renderer/components/Wallet/WalletDetails/WalletDetails.vue
@@ -9,6 +9,26 @@
       class="flex-1 overflow-y-auto"
     >
       <MenuTabItem
+        key="BackItem"
+        :label="$t('COMMON.BACK')"
+        :on-click="historyBack"
+      >
+        <div
+          slot="header"
+          class="WalletDetails__back-button flex items-center"
+        >
+          <SvgIcon
+            name="send"
+            view-box="0 0 8 8"
+          />
+          <span
+            class="text-bold ml-2 text-base"
+          >
+            {{ $t('COMMON.BACK') }}
+          </span>
+        </div>
+      </MenuTabItem>
+      <MenuTabItem
         v-for="tab in tabs"
         :key="tab.componentName"
         :label="tab.text"
@@ -137,6 +157,7 @@
 </template>
 
 <script>
+import electron from 'electron'
 import { at, clone } from 'lodash'
 /* eslint-disable vue/no-unused-components */
 import { WalletSelectDelegate } from '@/components/Wallet'
@@ -324,6 +345,15 @@ export default {
   },
 
   methods: {
+    historyBack () {
+      const webContents = electron.remote.getCurrentWindow().webContents
+      if (!webContents.canGoBack()) {
+        throw new Error('It is not possible to go back in history')
+      }
+
+      webContents.goBack()
+    },
+
     switchToTab (component) {
       this.currentTab = component
     },
@@ -444,5 +474,8 @@ export default {
 .WalletDetails__button:hover {
   transition: 0.5s;
   @apply .text-theme-voting-banner-button-text-hover .bg-theme-voting-banner-button-hover
+}
+.WalletDetails__back-button > svg {
+  transform: rotate(-135deg)
 }
 </style>
